misc: update snake_case normalizer documentation and rename configurator
