+++ docs/pages/serialization/common-transformers-examples.md
@@ -29,135 +29,9 @@ See [date time configurator chapter].
 
 ## Transforming property name to “snake_case”
 
-Depending on the conventions of the data format, it may be necessary to
-transform the case of the keys, for instance from “camelCase” to “snake_case”.
+See [snake_case configurator chapter].
 
-If this transformation is needed on every object, it can be done globally by
-using a global transformer, as shown in the example below:
-
-<details>
-<summary>Show code example — Global “snake_case” properties</summary>
-
-```php
-namespace My\App;
-
-final class CamelToSnakeCaseTransformer
-{
-    public function __invoke(object $object, callable $next): mixed
-    {
-        $result = $next();
-
-        if (! is_array($result)) {
-            return $result;
-        }
-
-        $snakeCased = [];
-
-        foreach ($result as $key => $value) {
-            $newKey = strtolower(preg_replace('/[A-Z]/', '_$0', lcfirst($key)));
-
-            $snakeCased[$newKey] = $value;
-        }
-
-        return $snakeCased;
-    }
-}
-
-(new \CuyZ\Valinor\NormalizerBuilder())
-    ->registerTransformer(new \My\App\CamelToSnakeCaseTransformer())
-    ->normalizer(\CuyZ\Valinor\Normalizer\Format::array())
-    ->normalize(
-         new \My\App\User(
-            name: 'John Doe',
-            emailAddress: 'john.doe@example.com', 
-            age: 42,
-            country: new Country(
-                name: 'France',
-                countryCode: 'FR',
-            ),
-        )
-    );
-
-// [
-//     'name' => 'John Doe',
-//     'email_address' => 'john.doe@example.com', // snake_case
-//     'age' => 42,
-//     'country' => [
-//         'name' => 'France',
-//         'country_code' => 'FR', // snake_case
-//     ],
-// ]
-```
-
-</details>
-
-For a more granular control, an attribute can be used to target specific
-objects, as shown in the example below:
-
-<details>
-<summary>Show code example — “snake_case” attribute</summary>
-
-```php
-namespace My\App;
-
-#[\CuyZ\Valinor\Normalizer\AsTransformer]
-#[\Attribute(\Attribute::TARGET_CLASS)]
-final class SnakeCaseProperties
-{
-    public function normalize(object $object, callable $next): array
-    {
-        $result = $next();
-
-        if (! is_array($result)) {
-            return $result;
-        }
-
-        $snakeCased = [];
-
-        foreach ($result as $key => $value) {
-            $newKey = strtolower(preg_replace('/[A-Z]/', '_$0', lcfirst($key)));
-
-            $snakeCased[$newKey] = $value;
-        }
-
-        return $snakeCased;
-    }
-}
-
-#[\My\App\SnakeCaseProperties]
-final readonly class Country
-{
-    public function __construct(
-        public string $name,
-        public string $countryCode,
-    ) {}
-}
-
-(new \CuyZ\Valinor\NormalizerBuilder())
-    ->normalizer(\CuyZ\Valinor\Normalizer\Format::array())
-    ->normalize(
-        new \My\App\User(
-            name: 'John Doe',
-            emailAddress: 'john.doe@example.com',
-            age: 42,
-            country: new Country(
-                name: 'France',
-                countryCode: 'FR',
-            ),
-        )
-    );
-
-// [
-//     'name' => 'John Doe',
-//     'emailAddress' => 'john.doe@example.com', // camelCase
-//     'age' => 42,
-//     'country' => [
-//         'name' => 'France',
-//         'country_code' => 'FR', // snake_case
-//     ],
-// ]
-```
-</details>
+[snake_case configurator chapter]: use-provided-normalizer-configurators.md#converting-keys-to-snake_case
 
 ## Ignoring properties
